clean up and moving layer title
