+++ packages/geoview-core/src/core/components/common/index.ts
@@ -1,5 +1,3 @@
-export * from './close-button';
-export * from './enlarge-button';
 export * from './focus-trap-element';
 export * from './responsive-grid';
 export * from './layer-list';
