+++ packages/geoview-core/src/core/components/layers/layers-panel.tsx
@@ -1,5 +1,6 @@
 import { useRef } from 'react';
 import Markdown from 'markdown-to-jsx';
+import { useTheme } from '@mui/material';
 import { Box, Paper } from '@/ui';
 import { useLayerDisplayState, useSelectedLayer } from '@/core/stores/store-interface-and-intial-values/layer-state';
 import { LayersToolbar } from './layers-toolbar';
@@ -8,8 +9,10 @@ import { LeftPanel } from './left-panel/left-panel';
 import { logger } from '@/core/utils/logger';
 import { useAppGuide } from '@/core/stores/store-interface-and-intial-values/app-state';
 import { ResponsiveGridLayout, ResponsiveGridLayoutExposedMethods } from '../common/responsive-grid-layout';
+import { Typography } from '@/ui/typography/typography';
 
 export function LayersPanel(): JSX.Element {
+  const theme = useTheme();
   // Log
   logger.logTraceRender('components/layers/layers-panel');
 
@@ -94,11 +97,28 @@ export function LayersPanel(): JSX.Element {
     return null;
   };
 
+  const layerTitle = (): JSX.Element => {
+    return (
+      <Typography
+        sx={{
+          fontSize: theme.palette.geoViewFontSize.lg,
+          fontWeight: '600',
+          marginTop: '12px',
+          [theme.breakpoints.up('md')]: { display: 'none' },
+        }}
+        component="div"
+      >
+        {selectedLayer?.layerName ?? ''}
+      </Typography>
+    );
+  };
+
   return (
     <ResponsiveGridLayout
       ref={responsiveLayoutRef}
       leftTop={<LayersToolbar />}
       leftMain={leftPanel()}
+      rightTop={layerTitle()}
       rightMain={rightPanel()}
       fullWidth={false}
     />
